Add open/save to HDF5 ability to mne.Report
This allows you to save `mne.Report` objects in such a way that they can
be read back later. This in turn allows you to incrementally add figures
to a `Report` object across scripts.

The `mne.open_report` function reads back a previously saved
`mne.Report` object, or creates a new one if it doesn't exist yet. In
this way, it mimics the Python buildin `open` function. You can use it
like this:

    report = mne.read_report('report.h5')
    report.add_figures_to_section(...)
    report.save('report.h5')

Or with a context manager like this:

    with mne.open_report('report.h5') as report:
    	report.add_figures_to_section(...)
    # report is saved upon exiting the context block

With the content manager, you can specify that the report should be
saved both as HDF5 and HTML:

    with mne.open_report('report.h5', 'report.html') as report:
    	report.add_figures_to_section(...)
    # report is saved upon exiting the context block

In this manner, the HTML file can be incrementally updated.